Refactor things to unixMain + setup fork/exec
