+++ library/process/src/darwinMain/kotlin/io/matthewnelson/kmp/process/internal/spawn/PosixSpawnFileActions.kt
@@ -15,21 +15,38 @@
  **/
 @file:Suppress("EXPECT_ACTUAL_CLASSIFIERS_ARE_IN_BETA_WARNING")
 
-package io.matthewnelson.kmp.process.internal
+package io.matthewnelson.kmp.process.internal.spawn
 
 import io.matthewnelson.kmp.file.IOException
+import io.matthewnelson.kmp.process.internal.check
 import kotlinx.cinterop.CValuesRef
 import kotlinx.cinterop.ExperimentalForeignApi
 import kotlinx.cinterop.MemScope
+import kotlinx.cinterop.alloc
+import kotlinx.cinterop.ptr
 
 @OptIn(ExperimentalForeignApi::class)
-internal expect value class PosixSpawnAttrs private constructor(
+internal actual value class PosixSpawnFileActions private actual constructor(
     private val _ref: CValuesRef<*>,
 ) {
 
-    internal companion object {
+    internal val ref: CValuesRef<posix_spawnattr_tVar> get() {
+        @Suppress("UNCHECKED_CAST")
+        return _ref as CValuesRef<posix_spawn_file_actions_tVar>
+    }
+
+    internal actual fun adddup2(fd: Int, newFd: Int): Int {
+        return posix_spawn_file_actions_adddup2(ref, fd, newFd)
+    }
+
+    internal actual companion object {
 
         @Throws(IOException::class)
-        internal fun MemScope.posixSpawnAttrInit(): PosixSpawnAttrs
+        internal actual fun MemScope.posixSpawnFileActionsInit(): PosixSpawnFileActions {
+            val fileActions = alloc<posix_spawn_file_actions_tVar>()
+            posix_spawn_file_actions_init(fileActions.ptr).check()
+            defer { posix_spawn_file_actions_destroy(fileActions.ptr) }
+            return PosixSpawnFileActions(fileActions.ptr)
+        }
     }
 }
