+++ library/process/src/linuxTest/kotlin/io/matthewnelson/kmp/process/internal/GnuLibcVersionUnitTest.kt
@@ -15,6 +15,7 @@
  **/
 package io.matthewnelson.kmp.process.internal
 
+import io.matthewnelson.kmp.process.internal.spawn.GnuLibcVersion
 import kotlin.test.*
 
 class GnuLibcVersionUnitTest {
