+++ library/process/src/nativeInterop/cinterop/glibc_version.def
@@ -1,3 +1,3 @@
-package = io.matthewnelson.kmp.process.internal
+package = io.matthewnelson.kmp.process.internal.spawn
 headers = gnu/libc-version.h
 headerFilter = gnu/libc-version.h
