+++ library/process/src/nativeInterop/cinterop/spawn.def
@@ -1,3 +1,3 @@
-package = io.matthewnelson.kmp.process.internal
+package = io.matthewnelson.kmp.process.internal.spawn
 headers = spawn.h
 headerFilter = spawn.h
